+++ src/test/clojure/convex_web/command_test.clj
@@ -146,6 +146,22 @@
               :convex-web.result/value "{:description \"Increments the given Long value by 1.\",:signature [{:return Long,:params [num]}],:errors {:CAST \"If the actor argument is not castable to Long.\"},:examples [{:code \"(inc 10)\"}]}"}
             (select-keys result [:convex-web.result/type
                                  :convex-web.result/value])))))
+
+  (testing "Metadata"
+    (let [execute (fn [source]
+                    (c/execute system {::c/id (java.util.UUID/randomUUID)
+                                       ::c/timestamp 1
+                                       ::c/mode :convex-web.command.mode/query
+                                       ::c/signer {:convex-web.account/address 9}
+                                       ::c/query
+                                       {:convex-web.query/source source
+                                        :convex-web.query/language :convex-lisp}}))
+
+          result-metadata (fn [command]
+                            (get-in command [:convex-web.command/result :convex-web.result/metadata]))]
+
+      (is (= true (contains? (result-metadata (execute "inc")) :doc)))
+      (is (= true (contains? (result-metadata (execute "loop")) :doc)))))
   
   (testing "Syntax error"
     (let [command (c/execute system {::c/id (java.util.UUID/randomUUID)